backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
gunicorn==21.2.0
packaging==25.0
psycopg2-binary==2.9.9
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.6.0
